refactor(app): extract file picker handling from inputs
